Optimize GET /imports.json [refs #89423736]
I removed some unit tests in the process. They were too hard to
maintain, and they tested stuff that was way too specific. Obviously, we
need some refactoring here.

Now GET /imports.json takes 4ms on my dev machine. It used to take 50ms.

+++ app/views/ImportJob/index.json.scala
@@ -5,11 +5,10 @@ import play.api.libs.json.JsValue
 import play.api.libs.json.Json.toJson
 import play.api.mvc.RequestHeader
 
-import org.overviewproject.tree.orm.{ DocumentSet, DocumentSetCreationJob }
-
+import org.overviewproject.models.{ DocumentSet, DocumentSetCreationJob }
 
 object index {
-  def apply(tuples: Iterable[(DocumentSetCreationJob, DocumentSet, Long)])(implicit lang: Lang, request: RequestHeader) : JsValue = {
+  def apply(tuples: Iterable[(DocumentSetCreationJob, DocumentSet, Int)])(implicit lang: Lang, request: RequestHeader) : JsValue = {
     val items : Iterable[JsValue] = tuples.map(Function.tupled(show.apply))
     toJson(items)
   }
